% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscalibration_test.R
\name{miscalibration_test}
\alias{miscalibration_test}
\alias{miscalibration_test.reliabilitydiag}
\alias{miscalibration_test.numeric}
\title{Miscalibration Test}
\usage{
miscalibration_test(x, ...)

\method{miscalibration_test}{reliabilitydiag}(x, ...)

\method{miscalibration_test}{numeric}(x, y, ...)
}
\arguments{
\item{x}{an \R object inheriting from \code{'reliabilitydiag'} or a numeric
vector of probability predictions taking values in [0, 1].}

\item{...}{further arguments to be passed to or from methods.}

\item{y}{a numeric vector of binary response values in \{0, 1\} to be
predicted.}
}
\value{
returns a \code{'tibble'} with entries
\tabular{ll}{
   \code{forecast} \tab the name of the prediction method. \cr
   \code{miscalibration} \tab the miscalibration statistic
   (see \code{\link{summary.reliabilitydiag}}). \cr
   \code{pvalue} \tab the pvalue.
 }
}
\description{
(experimental)
}
