% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.reliabilitydiag}
\alias{plot.reliabilitydiag}
\alias{autoplot.reliabilitydiag}
\title{Plotting reliability diagram objects}
\usage{
\method{plot}{reliabilitydiag}(x, ...)

\method{autoplot}{reliabilitydiag}(
  object,
  ...,
  colour = "red",
  params_histogram = list(yscale = 0.2, colour = "black", fill = NA),
  params_ribbon = list(fill = "blue", alpha = 0.15),
  params_diagonal = list(size = 0.3),
  params_CEPline = list(size = 0.2, colour = colour),
  params_CEPsegment = list(size = 2, colour = colour),
  params_CEPpoint = list(size = 2, colour = colour, shape = 16)
)
}
\arguments{
\item{x}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{...}{further arguments to be passed to or from methods.}

\item{object}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{colour}{a colour to be used in the CEP layers.}

\item{params_histogram}{a list of arguments for
\code{ggplot2::geom_histogram}; ignored if \code{length(x) > 1}.}

\item{params_ribbon}{a list of arguments for \code{ggplot2::geom_ribbon};
ignored if \code{length(x) > 1}.}

\item{params_diagonal}{a list of arguments for \code{ggplot2::geom_line}.}

\item{params_CEPline}{a list of arguments for \code{ggplot2::geom_line}.}

\item{params_CEPsegment}{a list of arguments for
\code{ggplot2::geom_segment}; only used for \code{"continuous"} predictions
and ignored if \code{length(x) > 1}.}

\item{params_CEPpoint}{a list of arguments for \code{ggplot2::geom_point};
only used for \code{"discrete"} predictions and ignored if
\code{length(x) > 1}.}
}
\description{
Plotting reliability diagram objects
}
