% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.reliabilitydiag}
\alias{print.reliabilitydiag}
\alias{print.summary.reliabilitydiag}
\title{Printing reliability diagram objects}
\usage{
\method{print}{reliabilitydiag}(x, ...)

\method{print}{summary.reliabilitydiag}(x, ...)
}
\arguments{
\item{x}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Printing reliability diagram objects
}
