% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilitydiag.R
\name{reliabilitydiag}
\alias{reliabilitydiag}
\alias{reliabilitydiag0}
\title{Reliability diagram object}
\usage{
reliabilitydiag(
  ...,
  y = NULL,
  r = NULL,
  tol = sqrt(.Machine$double.eps),
  xtype = NULL,
  xvalues = NULL,
  xnames = NULL,
  region.level = 0.9,
  region.method = NULL,
  region.position = "diagonal",
  n.boot = 100
)

reliabilitydiag0(y)
}
\arguments{
\item{...}{objects to be coerced to \code{'reliabilitydiag'} and concatenated}

\item{y}{a numeric vector of binary response values in \{0, 1\} to be
predicted.}

\item{r}{an object inheriting from the class \code{'reliabilitydiag'};
alternative to \code{y}.}

\item{tol}{accuracy when comparing \code{y} in \code{'reliabilitydiag'}
objects.}

\item{xtype}{a string specifying whether the prediction values should be
treated as \code{"continuous"} or \code{"discrete"}.}

\item{xvalues}{a numeric vector of possible prediction values;
values in \code{x} are rounded to the nearest value in \code{xvalues} and
\code{xtype} is set to \code{"discrete"}.}

\item{xnames}{a character vector of prediction names.}

\item{region.level}{a value in (0, 1) specifying the level at which
consistency or confidence regions are calculated.}

\item{region.method}{a string specifying whether \code{"resampling"},
\code{"continuous_asymptotics"}, or \code{"discrete_asymptotics"} are used
to calculate consistency/confidence regions.}

\item{region.position}{a string specifying whether consistency regions
around the \code{"diagonal"} or confidence regions around the
\code{"estimate"} are calculated.}

\item{n.boot}{the number of bootstrap samples when
\code{region.method == "resampling"}.}
}
\value{
\code{reliabilitydiag} returns a \code{'reliabilitydiag'} object,
 which is a named list-type vector class with attribute
 \tabular{ll}{
   \code{y} \tab a numeric vector of response values to be predicted.
 }
 Each entry of a \code{'reliabilitydiag'} object is a list
 with the following components:
 \tabular{ll}{
   \code{cases} \tab a \code{'tibble'}.\cr
   \code{bins} \tab a \code{'tibble'}.\cr
   \code{regions} \tab a \code{'tibble'}.\cr
   \code{xinfo} \tab a \code{'list'}.
 }

 \code{reliabilitydiag0} returns an empty \code{'reliabilitydiag'} object
 with attribute \code{y}.
}
\description{
Documentation of the \code{'reliabilitydiag'} object, and its constructors.
}
\details{
\code{reliabilitydiag} constructs and returns an object inheriting from the
class \code{'reliabilitydiag'}. It calls the
concatenation function described in \code{\link{c.reliabilitydiag}}, which
in turn calls the
coercion methods described in \code{\link{as.reliabilitydiag}}.

\code{reliabilitydiag0} constructs an empty \code{'reliabilitydiag'} object
from the response values.
}
\examples{
set.seed(42)

X <- runif(100)
Y <- rbinom(100, 1, X)
r <- reliabilitydiag(X, y = Y)
r

r0 <- reliabilitydiag0(Y)
identical(r, reliabilitydiag(X, r = r0))

X2 <- runif(100)
r1 <- reliabilitydiag(X2, r = r)
r2 <- reliabilitydiag(X2, r = r0)
identical(r1, r2)

}
\seealso{
\code{\link{c.reliabilitydiag}},
 \code{\link{[.reliabilitydiag}},
 \code{\link{plot.reliabilitydiag}}.
}
