% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_extract.R
\name{[.reliabilitydiag}
\alias{[.reliabilitydiag}
\title{Subsetting reliability diagram objects}
\usage{
\method{[}{reliabilitydiag}(x, i)
}
\arguments{
\item{x}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{i}{index specifying which elements to extract.}
}
\value{
an object inheriting from the class \code{'reliabilitydiag'}.
}
\description{
Subsetting reliability diagram objects
}
\examples{
data("precip_Niamey_2016", package = "reliabilitydiag")

r <- reliabilitydiag(
  precip_Niamey_2016[c("Logistic", "EMOS", "ENS", "EPC")],
  y = precip_Niamey_2016$obs
)
length(r)
r[1]
r["EMOS"]

}
\seealso{
\code{\link{c.reliabilitydiag}}.
}
