% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.reliabilitydiag}
\alias{summary.reliabilitydiag}
\title{Summary of reliability diagram objects}
\usage{
\method{summary}{reliabilitydiag}(object, ..., score = "brier")
}
\arguments{
\item{object}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{...}{further arguments to be passed to or from methods.}

\item{score}{currently only "brier"}
}
\description{
Summary of reliability diagram objects
}
