% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.reliabilitydiag}
\alias{plot.reliabilitydiag}
\alias{autoplot.reliabilitydiag}
\title{Plotting reliability diagram objects}
\usage{
\method{plot}{reliabilitydiag}(x, ...)

\method{autoplot}{reliabilitydiag}(
  object,
  ...,
  type = c("miscalibration", "discrimination"),
  colour = "red",
  params_histogram = NULL,
  params_ggMarginal = NULL,
  params_ribbon = NULL,
  params_diagonal = NULL,
  params_vsegment = NULL,
  params_hsegment = NULL,
  params_CEPline = NULL,
  params_CEPsegment = NULL,
  params_CEPpoint = NULL
)
}
\arguments{
\item{x}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{...}{further arguments to be passed to or from methods.}

\item{object}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{type}{one of \code{"miscalibration"}, \code{"discrimination"};
determines which
layers are added by default, including default parameter values.}

\item{colour}{a colour to be used to draw focus;
used for the CEP layers when \code{type} is \code{"miscalibration"},
and for the horizontal segment layer and CEP margin histogram
when \code{type} is \code{"discrimination"}.}

\item{params_histogram}{a list of arguments for
\code{ggplot2::geom_histogram}.}

\item{params_ggMarginal}{a list of arguments for \code{ggExtra::ggMarginal}.}

\item{params_ribbon}{a list of arguments for \code{ggplot2::geom_ribbon}.}

\item{params_diagonal}{a list of arguments for \code{ggplot2::geom_line}.}

\item{params_vsegment}{a list of arguments for \code{ggplot2::geom_segment}.}

\item{params_hsegment}{a list of arguments for \code{ggplot2::geom_segment}.}

\item{params_CEPline}{a list of arguments for \code{ggplot2::geom_line}.}

\item{params_CEPsegment}{a list of arguments for
\code{ggplot2::geom_segment}.}

\item{params_CEPpoint}{a list of arguments for \code{ggplot2::geom_point}.}
}
\description{
Plotting reliability diagram objects
}
