% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.reliabilitydiag}
\alias{print.reliabilitydiag}
\alias{print.summary.reliabilitydiag}
\title{Printing reliability diagram objects}
\usage{
\method{print}{reliabilitydiag}(x, ...)

\method{print}{summary.reliabilitydiag}(x, ...)
}
\arguments{
\item{x}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{...}{further arguments to be passed to or from methods;
in particular, these are passed to \code{\link{autoplot.reliabilitydiag}}
and \code{\link[tibble]{print.tbl_df}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Printing methods for \code{'reliabilitydiag'} and
\code{'summary.reliabilitydiag'} objects.
}
\details{
\code{print.reliabilitydiag} always sends a plot to the current graphics
device and prints a summary to the console.

\code{print.summary.reliabilitydiag} prints the summary output to the console.
}
\seealso{
\code{\link{autoplot.reliabilitydiag}},
  \code{\link{summary.reliabilitydiag}}
}
