% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilitydiag.R
\name{reliabilitydiag}
\alias{reliabilitydiag}
\alias{reliabilitydiag0}
\title{Reliability diagram object}
\usage{
reliabilitydiag(
  ...,
  y = NULL,
  r = NULL,
  tol = sqrt(.Machine$double.eps),
  xtype = NULL,
  xvalues = NULL,
  xnames = NULL,
  region.level = 0.9,
  region.method = NULL,
  region.position = "diagonal",
  n.boot = 100
)

reliabilitydiag0(y)
}
\arguments{
\item{...}{objects to be coerced to \code{'reliabilitydiag'} and concatenated}

\item{y}{a numeric vector of binary response values in \{0, 1\} to be
predicted.}

\item{r}{an object inheriting from the class \code{'reliabilitydiag'};
alternative to \code{y}.}

\item{tol}{accuracy when comparing \code{y} in \code{'reliabilitydiag'}
objects.}

\item{xtype}{a string specifying whether the prediction values should be
treated as \code{"continuous"} or \code{"discrete"}.}

\item{xvalues}{a numeric vector of possible prediction values;
values in \code{x} are rounded to the nearest value in \code{xvalues} and
\code{xtype} is set to \code{"discrete"}.}

\item{xnames}{a character vector of prediction names.}

\item{region.level}{a value in (0, 1) specifying the level at which
consistency or confidence regions are calculated.}

\item{region.method}{a string specifying whether \code{"resampling"},
\code{"continuous_asymptotics"}, or \code{"discrete_asymptotics"} are used
to calculate consistency/confidence regions.}

\item{region.position}{a string specifying whether consistency regions
around the \code{"diagonal"} or confidence regions around the
\code{"estimate"} are calculated.}

\item{n.boot}{the number of bootstrap samples when
\code{region.method == "resampling"}.}
}
\value{
\code{reliabilitydiag} returns a \code{'reliabilitydiag'} object,
 which is a named list-type vector class with the attribute
 \code{y} containing the values supplied to the input argument \code{y},
 that is, the numeric vector of response values to be predicted.
 The length is given by the number of prediction methods detected from the
 supplied objects.

 \code{reliabilitydiag0} returns an empty \code{'reliabilitydiag'} object
 with attribute \code{y}.

 Each entry of a \code{'reliabilitydiag'} object
 (corresponding to a single prediction method)
 is itself a list with the following entries
 \tabular{ll}{
   \code{cases} \tab a tibble of all predictions and observations.\cr
   \code{bins} \tab a tibble of the characteristics of the PAV induced bins.
     \cr
   \code{regions} \tab a tibble with lower and upper bounds of the pointwise
     consistency/confidence regions.\cr
   \code{xinfo} \tab a list of characteristics of \code{x}.
 }

 Each \code{cases} tibble comprises the forecast-observation pairs of the
 given prediction method. It is arranged in increasing order of
 \code{x} and has columns
 \tabular{ll}{
   \code{case_id} \tab an ID based on the original order of the predictions
     and observations.\cr
   \code{x} \tab an original prediction (increasing order).\cr
   \code{y} \tab an observation, corresponding to \code{x}.\cr
   \code{bin_id} \tab an ID for the PAV-recalibration induced bins.
 }

 Each \code{bins} tibble contains PAV-recalibration information, and has
 columns
 \tabular{ll}{
   \code{bin_id} \tab as in \code{cases}, with any ID only appearing
     once.\cr
   \code{n} \tab the number of predictions with a given \code{bin_id}.\cr
   \code{x_min} \tab the smallest value of the predictions with the given
     \code{bin_id}.\cr
   \code{x_max} \tab the largest value of the predictions with the given
     \code{bin_id}.\cr
   \code{CEP_pav} \tab the unique PAV-recalibrated prediction
     corresponding to \code{bin_id}.
 }

 Each \code{regions} tibble contains the uncertainty quantification
 information, and has columns
 \tabular{ll}{
   \code{x} \tab an original prediction, with any value only appearing
     once.\cr
   \code{lower} \tab the lower bound of the consistency/confidence
     region at \code{x}.\cr
   \code{upper} \tab the upper bound of the consistency/confidence
     region \code{x}.\cr
   \code{n} \tab the number of predictions with a value of \code{x}.\cr
   \code{level} \tab the level of the consistency/confidence regions. \cr
   \code{method} \tab the method used to calculate the
     consistency/confidence region.\cr
   \code{position} \tab \code{"diagonal"} for a consistency region, and
     \code{"estimate"} for a confidence region.
 }

 Each \code{xinfo} list has entries
 \tabular{ll}{
   \code{type} \tab the type of predictions, either \code{"discrete"}
       or \code{"continuous"}.\cr
   \code{values} \tab the values supplied to \code{xvalues}.
 }
}
\description{
Documentation of the \code{'reliabilitydiag'} object, and its constructors.
}
\details{
\code{reliabilitydiag} constructs and returns an object inheriting from the
class \code{'reliabilitydiag'}.
Each object passed via \code{...} is
coerced by the methods described in \code{\link{as.reliabilitydiag}},
and then concatenated by \code{\link{c.reliabilitydiag}}.

\code{reliabilitydiag0} constructs an empty \code{'reliabilitydiag'} object
from the response values.
}
\examples{
set.seed(42)

X <- runif(100)
Y <- rbinom(100, 1, X)
r <- reliabilitydiag(X, y = Y)
r

# no consistency/confidence regions
r1 <- reliabilitydiag(X, y = Y, region.level = NA)
r1

# specify predictions via existing reliabilitydiag
r0 <- reliabilitydiag0(Y)
identical(r1, reliabilitydiag(X, r = r0, region.level = NA))

# only observation information is used from existing reliabilitydiag
X2 <- runif(100)
r2 <- reliabilitydiag(X2, r = r, region.level = NA)
r3 <- reliabilitydiag(X2, r = r0, region.level = NA)
identical(r2, r3)

}
\seealso{
\code{\link{c.reliabilitydiag}},
 \code{\link{[.reliabilitydiag}},
 \code{\link{plot.reliabilitydiag}}.
}
