% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_extract.R
\name{[.reliabilitydiag}
\alias{[.reliabilitydiag}
\title{Subsetting reliability diagram objects}
\usage{
\method{[}{reliabilitydiag}(x, i)
}
\arguments{
\item{x}{an object inheriting from the class \code{'reliabilitydiag'}.}

\item{i}{index specifying which elements to extract.}
}
\value{
an object inheriting from the class \code{'reliabilitydiag'}.
}
\description{
Subsetting reliability diagram objects
}
\examples{
set.seed(42)

X1 <- runif(100)
X2 <- runif(100)
Y <- rbinom(100, 1, X1)
r <- reliabilitydiag(X1, X2, y = Y)
length(r)
r[1]

}
\seealso{
\code{\link{c.reliabilitydiag}}.
}
